+++ tests/end-to-end/api/test_api_endpoint.py
@@ -6,14 +6,16 @@ import sys
 import unittest
 
 
-# Depends on the provided tests first passing
-# user_login
-class TestDataFedPythonAPIContext(unittest.TestCase):
+# Depends on the success of tests:
+# 1. login
+class TestDataFedPythonAPIEndpoint(unittest.TestCase):
     def setUp(self):
         path_of_file = os.path.abspath(__file__)
         current_folder = os.path.dirname(path_of_file)
         path_to_python_datafed_module = os.path.normpath(
             current_folder
+            + os.sep
+            + ".."
             + os.sep
             + ".."
             + os.sep
@@ -44,13 +46,13 @@ class TestDataFedPythonAPIContext(unittest.TestCase):
 
         self._df_api = API(opts)
 
-        self._username = "datafed99"
-        password = os.environ.get("DATAFED_USER99_PASSWORD")
+        username = "datafed89"
+        password = os.environ.get("DATAFED_USER89_PASSWORD")
 
         count = 0
         while True:
             try:
-                self._df_api.loginByPassword(self._username, password)
+                self._df_api.loginByPassword(username, password)
                 break
             except BaseException:
                 pass
@@ -58,16 +60,28 @@ class TestDataFedPythonAPIContext(unittest.TestCase):
             # Try three times to authenticate
             assert count < 3
 
-    def test_context(self):
-        print("Running getContext()")
-        self._df_api.getContext()
-        self.assertEqual(self._df_api.getContext(), f"u/{self._username}")
+    def test_endpoint_set_and_default(self):
+        endpoint = os.environ.get("DATAFED_USER89_GLOBUS_UUID")
+        if endpoint is None:
+            self.fail(
+                "Cannot run end-to-end tests with Python CLI requires setting "
+                " env variable DATAFED_REPO_ENDPOINT_UUID so that we know what to"
+                " set the default endpoint to. This should be the same endpoint"
+                " that the users have an allocation on... users datafed89 and"
+                " datafed99"
+            )
+
+        if not self._df_api.endpointDefaultGet():
+            self._df_api.endpointDefaultSet(endpoint)
+
+        new_endpoint = self._df_api.endpointDefaultGet()
+        self.assertEqual(endpoint, new_endpoint)
 
 
 if __name__ == "__main__":
     suite = unittest.TestSuite()
     # Add them in the order they should be executed
-    suite.addTest(TestDataFedPythonAPIContext("test_context"))
+    suite.addTest(TestDataFedPythonAPIEndpoint("test_endpoint_set_and_default"))
     runner = unittest.TextTestRunner()
     result = runner.run(suite)
     # wasSuccessful() return True which is not 0