+++ tests/end-to-end/api/test_api_repo.py
@@ -19,6 +19,8 @@ class TestDataFedPythonAPIRepo(unittest.TestCase):
             + os.sep
             + ".."
             + os.sep
+            + ".."
+            + os.sep
             + "python/datafed_pkg"
         )
         sys.path.insert(0, path_to_python_datafed_module)